+++ src/package-extensions.js
@@ -4,7 +4,7 @@ import assert from "node:assert";
 import fs from "node:fs/promises";
 import path from "node:path";
 import { sortExtensionsToml } from "./lib/extensions-toml.js";
-import { fileExists, readTomlFile } from "./lib/fs.js";
+import { fileExists, readTomlFile, readExtensionFiles } from "./lib/fs.js";
 import {
   checkoutGitSubmodule,
   readGitmodules,
@@ -140,7 +140,8 @@ async function packageExtension(
   extensionVersion,
   shouldPublish,
 ) {
-  validateLicense(extensionPath);
+  const extensionFiles = await readExtensionFiles(extensionPath);
+  validateLicense(extensionFiles);
 
   const outputDir = "output";
